Move v1 Evo* TypedDicts to typedefs.py
